refactor: split ledger
