+++ ledger/ledger.go
@@ -3,18 +3,13 @@ package ledger
 import (
 	"context"
 	"encoding/binary"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"math/rand"
-	"sort"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/dgraph-io/badger"
-	"github.com/dgraph-io/badger/pb"
 	"github.com/qlcchain/go-qlc/common"
 	"github.com/qlcchain/go-qlc/common/event"
 	"github.com/qlcchain/go-qlc/common/types"
@@ -245,6 +240,59 @@ func (l *Ledger) DBStore() db.Store {
 	return l.Store
 }
 
+// BatchUpdate MUST pass the same txn
+func (l *Ledger) BatchUpdate(fn func(txn db.StoreTxn) error) error {
+	txn := l.Store.NewTransaction(true)
+	//logger.Debugf("BatchUpdate NewTransaction %p", txn)
+	defer func() {
+		//logger.Debugf("BatchUpdate Discard %p", txn)
+		txn.Discard()
+	}()
+
+	if err := fn(txn); err != nil {
+		return err
+	}
+	return txn.Commit(nil)
+}
+
+// BatchView MUST pass the same txn
+func (l *Ledger) BatchView(fn func(txn db.StoreTxn) error) error {
+	txn := l.Store.NewTransaction(false)
+	//logger.Debugf("BatchView NewTransaction %p", txn)
+	defer func() {
+		//logger.Debugf("BatchView Discard %p", txn)
+		txn.Discard()
+	}()
+
+	if err := fn(txn); err != nil {
+		return err
+	}
+	return txn.Commit(nil)
+}
+
+//getTxn get txn by `update` mode
+func (l *Ledger) getTxn(update bool, txns ...db.StoreTxn) (db.StoreTxn, bool) {
+	if len(txns) > 0 {
+		//logger.Debugf("getTxn %p", txns[0])
+		return txns[0], false
+	} else {
+		txn := l.Store.NewTransaction(update)
+		//logger.Debugf("getTxn new %p", txn)
+		return txn, true
+	}
+}
+
+// releaseTxn commit change and close txn
+func (l *Ledger) releaseTxn(txn db.StoreTxn, flag bool) {
+	if flag {
+		err := txn.Commit(nil)
+		if err != nil {
+			l.logger.Error(err)
+		}
+		txn.Discard()
+	}
+}
+
 func getKeyOfParts(t byte, partList ...interface{}) ([]byte, error) {
 	var buffer = []byte{t}
 	for _, part := range partList {
@@ -288,1449 +336,6 @@ func getKeyOfParts(t byte, partList ...interface{}) ([]byte, error) {
 	return buffer, nil
 }
 
-func (l *Ledger) AddStateBlock(value *types.StateBlock, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-
-	k, err := getKeyOfParts(idPrefixBlock, value.GetHash())
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrBlockExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	if err := txn.Set(k, v); err != nil {
-		return err
-	}
-
-	if b, err := l.HasBlockCache(value.GetHash()); b && err == nil {
-		if err := l.DeleteBlockCache(value.GetHash(), txn); err != nil {
-			return fmt.Errorf("delete block cache error: %s", err)
-		}
-	}
-	if err := addChild(value, txn); err != nil {
-		return fmt.Errorf("add block child error: %s", err)
-	}
-	if err := addLink(value, txn); err != nil {
-		return fmt.Errorf("add block link error: %s", err)
-	}
-	l.releaseTxn(txn, flag)
-	l.logger.Debug("publish addRelation,", value.GetHash())
-	l.EB.Publish(common.EventAddRelation, value)
-	return nil
-}
-
-func addChild(cBlock *types.StateBlock, txn db.StoreTxn) error {
-	pHash := cBlock.Parent()
-	cHash := cBlock.GetHash()
-	if !common.IsGenesisBlock(cBlock) && pHash != types.ZeroHash && !cBlock.IsOpen() {
-		// is parent block existed
-		kCache, err := getKeyOfParts(idPrefixBlockCache, pHash)
-		if err != nil {
-			return err
-		}
-		err = txn.Get(kCache, func(v []byte, b byte) error {
-			return nil
-		})
-		if err != nil {
-			kConfirmed, err := getKeyOfParts(idPrefixBlock, pHash)
-			if err != nil {
-				return err
-			}
-			err = txn.Get(kConfirmed, func(v []byte, b byte) error {
-				return nil
-			})
-			if err != nil {
-				return fmt.Errorf("%s can not find parent %s", cHash.String(), pHash.String())
-			}
-		}
-
-		// is parent have used
-		k, err := getKeyOfParts(idPrefixChild, pHash)
-		if err != nil {
-			return err
-		}
-
-		err = txn.Get(k, func(val []byte, b byte) error {
-			return nil
-		})
-		if err == nil {
-			return fmt.Errorf("%s already have child ", pHash.String())
-		}
-
-		// add new relationship
-		v, err := cHash.MarshalMsg(nil)
-		if err != nil {
-			return err
-		}
-		if err := txn.Set(k, v); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func addLink(block *types.StateBlock, txn db.StoreTxn) error {
-	if block.GetType() == types.Open || block.GetType() == types.Receive || block.GetType() == types.ContractReward {
-		k, err := getKeyOfParts(idPrefixLink, block.GetLink())
-		if err != nil {
-			return err
-		}
-		h := block.GetHash()
-		v, err := h.MarshalMsg(nil)
-		if err != nil {
-			return err
-		}
-		if err := txn.Set(k, v); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (l *Ledger) GetStateBlock(key types.Hash, txns ...db.StoreTxn) (*types.StateBlock, error) {
-	if blkCache, err := l.GetBlockCache(key); err == nil {
-		return blkCache, nil
-	}
-
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlock, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.StateBlock)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrBlockNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetStateBlockConfirmed(key types.Hash, txns ...db.StoreTxn) (*types.StateBlock, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlock, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.StateBlock)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrBlockNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetStateBlocks(fn func(*types.StateBlock) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	errStr := make([]string, 0)
-	err := txn.Iterator(idPrefixBlock, func(key []byte, val []byte, b byte) error {
-		blk := new(types.StateBlock)
-		if err := blk.Deserialize(val); err != nil {
-			l.logger.Errorf("deserialize block error: %s", err)
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		if err := fn(blk); err != nil {
-			l.logger.Errorf("process block error: %s", err)
-			errStr = append(errStr, err.Error())
-		}
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-	if len(errStr) != 0 {
-		return errors.New(strings.Join(errStr, ", "))
-	}
-	return nil
-}
-
-func (l *Ledger) DeleteStateBlock(key types.Hash, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-
-	k, err := getKeyOfParts(idPrefixBlock, key)
-	if err != nil {
-		return err
-	}
-
-	blk := new(types.StateBlock)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := blk.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	if err := txn.Delete(k); err != nil {
-		return err
-	}
-
-	if err := l.deleteChild(blk, txn); err != nil {
-		return fmt.Errorf("delete child error: %s", err)
-	}
-	if err := l.deleteLink(blk, txn); err != nil {
-		return fmt.Errorf("delete link error: %s", err)
-	}
-
-	l.releaseTxn(txn, flag)
-	l.logger.Info("publish deleteRelation,", key.String())
-	l.EB.Publish(common.EventDeleteRelation, key)
-	return nil
-}
-
-func (l *Ledger) deleteChild(blk *types.StateBlock, txn db.StoreTxn) error {
-	pHash := blk.Parent()
-	if !pHash.IsZero() {
-
-		k, err := getKeyOfParts(idPrefixChild, pHash)
-		if err != nil {
-			return err
-		}
-		if err := txn.Delete(k); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (l *Ledger) deleteLink(blk *types.StateBlock, txn db.StoreTxn) error {
-	k, err := getKeyOfParts(idPrefixLink, blk.GetLink())
-	if err != nil {
-		return err
-	}
-	if err := txn.Delete(k); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) HasStateBlock(key types.Hash, txns ...db.StoreTxn) (bool, error) {
-	if exit, err := l.HasBlockCache(key); err == nil && exit {
-		return exit, nil
-	}
-
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlock, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) HasStateBlockConfirmed(key types.Hash, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlock, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) CountStateBlocks(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Count([]byte{idPrefixBlock})
-}
-
-func (l *Ledger) GetRandomStateBlock(txns ...db.StoreTxn) (*types.StateBlock, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	c, err := l.CountStateBlocks()
-	if err != nil {
-		return nil, err
-	}
-	if c == 0 {
-		return nil, ErrStoreEmpty
-	}
-	blk := new(types.StateBlock)
-	errFound := errors.New("state block found")
-
-	for i := 0; i < 3; i++ {
-		index := rand.Int63n(int64(c))
-		var temp int64
-		err = txn.Iterator(idPrefixBlock, func(key []byte, val []byte, b byte) error {
-			if temp == index {
-				var b = new(types.StateBlock)
-				if err = b.Deserialize(val); err != nil {
-					return err
-				}
-				if !common.IsGenesisBlock(b) {
-					blk = b
-					return errFound
-				}
-			}
-			temp++
-			return nil
-		})
-		if err != nil && err != errFound {
-			return nil, err
-		}
-		if !blk.Token.IsZero() {
-			break
-		}
-	}
-	if blk.Token.IsZero() {
-		return nil, errors.New("state block not found")
-	}
-	return blk, nil
-}
-
-func (l *Ledger) AddSmartContractBlock(value *types.SmartContractBlock, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixSmartContractBlock, value.GetHash())
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrBlockExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) GetSmartContractBlock(key types.Hash, txns ...db.StoreTxn) (*types.SmartContractBlock, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixSmartContractBlock, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.SmartContractBlock)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrBlockNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetSmartContractBlocks(fn func(block *types.SmartContractBlock) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	errStr := make([]string, 0)
-	err := txn.Iterator(idPrefixSmartContractBlock, func(key []byte, val []byte, b byte) error {
-		blk := new(types.SmartContractBlock)
-		if err := blk.Deserialize(val); err != nil {
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		if err := fn(blk); err != nil {
-			errStr = append(errStr, err.Error())
-		}
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-	if len(errStr) != 0 {
-		return errors.New(strings.Join(errStr, ", "))
-	}
-	return nil
-}
-
-func (l *Ledger) HasSmartContractBlock(key types.Hash, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixSmartContractBlock, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) CountSmartContractBlocks(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-	return txn.Count([]byte{idPrefixSmartContractBlock})
-}
-
-func (l *Ledger) uncheckedKindToPrefix(kind types.UncheckedKind) byte {
-	switch kind {
-	case types.UncheckedKindPrevious:
-		return idPrefixUncheckedBlockPrevious
-	case types.UncheckedKindLink:
-		return idPrefixUncheckedBlockLink
-	case types.UncheckedKindTokenInfo:
-		return idPrefixUncheckedTokenInfo
-	default:
-		panic("bad unchecked block kind")
-	}
-}
-
-func (l *Ledger) AddUncheckedBlock(key types.Hash, value *types.StateBlock, kind types.UncheckedKind, sync types.SynchronizedKind, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(l.uncheckedKindToPrefix(kind), key)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrUncheckedBlockExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.SetWithMeta(k, v, byte(sync))
-}
-
-func (l *Ledger) GetUncheckedBlock(key types.Hash, kind types.UncheckedKind, txns ...db.StoreTxn) (*types.StateBlock, types.SynchronizedKind, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(l.uncheckedKindToPrefix(kind), key)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	value := new(types.StateBlock)
-	var sync types.SynchronizedKind
-	err = txn.Get(k, func(val []byte, b byte) (err error) {
-		if err = value.Deserialize(val); err != nil {
-			return err
-		}
-		sync = types.SynchronizedKind(b)
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, 0, ErrUncheckedBlockNotFound
-		}
-		return nil, 0, err
-	}
-	return value, sync, nil
-}
-
-func (l *Ledger) DeleteUncheckedBlock(key types.Hash, kind types.UncheckedKind, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(l.uncheckedKindToPrefix(kind), key)
-	if err != nil {
-		return err
-	}
-	return txn.Delete(k)
-}
-
-func (l *Ledger) HasUncheckedBlock(key types.Hash, kind types.UncheckedKind, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(l.uncheckedKindToPrefix(kind), key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) walkUncheckedBlocks(kind types.UncheckedKind, visit types.UncheckedBlockWalkFunc, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	errStr := make([]string, 0)
-	prefix := l.uncheckedKindToPrefix(kind)
-	err := txn.Iterator(prefix, func(key []byte, val []byte, b byte) error {
-		blk := new(types.StateBlock)
-		if err := blk.Deserialize(val); err != nil {
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		h, err := types.BytesToHash(key[1:])
-		if err != nil {
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		if err := visit(blk, h, kind, types.SynchronizedKind(b)); err != nil {
-			l.logger.Error("visit error %s", err)
-			errStr = append(errStr, err.Error())
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	if len(errStr) != 0 {
-		return errors.New(strings.Join(errStr, ", "))
-	}
-	return nil
-}
-
-func (l *Ledger) WalkUncheckedBlocks(visit types.UncheckedBlockWalkFunc, txns ...db.StoreTxn) error {
-	if err := l.walkUncheckedBlocks(types.UncheckedKindPrevious, visit, txns...); err != nil {
-		return err
-	}
-
-	if err := l.walkUncheckedBlocks(types.UncheckedKindLink, visit, txns...); err != nil {
-		return err
-	}
-
-	return l.walkUncheckedBlocks(types.UncheckedKindTokenInfo, visit, txns...)
-}
-
-func (l *Ledger) CountUncheckedBlocks(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	var count uint64
-	count, err := txn.Count([]byte{idPrefixUncheckedBlockLink})
-	if err != nil {
-		return 0, err
-	}
-
-	count2, err := txn.Count([]byte{idPrefixUncheckedBlockPrevious})
-	if err != nil {
-		return 0, err
-	}
-
-	return count + count2, nil
-}
-
-func (l *Ledger) AddAccountMeta(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrAccountExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) GetAccountMeta(key types.Address, txns ...db.StoreTxn) (*types.AccountMeta, error) {
-	am, err := l.GetAccountMetaCache(key)
-	if err != nil {
-		am = nil
-	}
-
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, key)
-	if err != nil {
-		return nil, err
-	}
-	meta := new(types.AccountMeta)
-	err = txn.Get(k, func(v []byte, b byte) (err error) {
-		if err = meta.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-
-	if err != nil {
-		meta = nil
-	}
-	if am != nil && meta == nil {
-		return am, nil
-	}
-	if am == nil && meta != nil {
-		return meta, nil
-	}
-	if am != nil && meta != nil {
-		for _, v := range meta.Tokens {
-			temp := am.Token(v.Type)
-			if temp != nil {
-				if temp.BlockCount < v.BlockCount {
-					temp.BlockCount = v.BlockCount
-					temp.Type = v.Type
-					temp.BelongTo = v.BelongTo
-					temp.Header = v.Header
-					temp.Balance = v.Balance
-					temp.Modified = v.Modified
-					temp.OpenBlock = v.OpenBlock
-					temp.Representative = v.Representative
-				}
-			} else {
-				am.Tokens = append(am.Tokens, v)
-			}
-		}
-		return am, nil
-	}
-	return nil, ErrAccountNotFound
-}
-
-func (l *Ledger) GetAccountMetaConfirmed(key types.Address, txns ...db.StoreTxn) (*types.AccountMeta, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.AccountMeta)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrAccountNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetAccountMetas(fn func(am *types.AccountMeta) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixAccount, func(key []byte, val []byte, b byte) error {
-		am := new(types.AccountMeta)
-		if err := am.Deserialize(val); err != nil {
-			return err
-		}
-		if err := fn(am); err != nil {
-			return err
-		}
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) CountAccountMetas(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Count([]byte{idPrefixAccount})
-}
-
-func (l *Ledger) UpdateAccountMeta(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return ErrAccountNotFound
-		}
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) AddOrUpdateAccountMeta(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) DeleteAccountMeta(key types.Address, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, key)
-	if err != nil {
-		return err
-	}
-
-	return txn.Delete(k)
-}
-
-func (l *Ledger) HasAccountMeta(key types.Address, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixAccount, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) AddTokenMeta(address types.Address, meta *types.TokenMeta, txns ...db.StoreTxn) error {
-	am, err := l.GetAccountMeta(address, txns...)
-	if err != nil {
-		return err
-	}
-
-	if am.Token(meta.Type) != nil {
-		return ErrTokenExists
-	}
-
-	am.Tokens = append(am.Tokens, meta)
-	return l.UpdateAccountMeta(am, txns...)
-}
-
-func (l *Ledger) GetTokenMeta(address types.Address, tokenType types.Hash, txns ...db.StoreTxn) (*types.TokenMeta, error) {
-	am, err := l.GetAccountMeta(address, txns...)
-	if err != nil {
-		return nil, err
-	}
-
-	tm := am.Token(tokenType)
-	if tm == nil {
-		return nil, ErrTokenNotFound
-	}
-
-	return tm, nil
-}
-
-func (l *Ledger) GetTokenMetaConfirmed(address types.Address, tokenType types.Hash, txns ...db.StoreTxn) (*types.TokenMeta, error) {
-	am, err := l.GetAccountMetaConfirmed(address, txns...)
-	if err != nil {
-		return nil, err
-	}
-
-	tm := am.Token(tokenType)
-	if tm == nil {
-		return nil, ErrTokenNotFound
-	}
-
-	return tm, nil
-}
-
-func (l *Ledger) UpdateTokenMeta(address types.Address, meta *types.TokenMeta, txns ...db.StoreTxn) error {
-	am, err := l.GetAccountMeta(address, txns...)
-	if err != nil {
-		return err
-	}
-
-	//tm := am.Token(meta.Type)
-	//
-	//if tm != nil {
-	//	tm = meta
-	//	return l.UpdateAccountMeta(am, txns...)
-	//}
-	//tokens := am.Tokens
-	for index, token := range am.Tokens {
-		if token.Type == meta.Type {
-			//am.Tokens = append(tokens[:index], tokens[index+1:]...)
-			//am.Tokens = append(am.Tokens, meta)
-			am.Tokens[index] = meta
-			return l.UpdateAccountMeta(am, txns...)
-		}
-	}
-	return ErrTokenNotFound
-}
-
-func (l *Ledger) AddOrUpdateTokenMeta(address types.Address, meta *types.TokenMeta, txns ...db.StoreTxn) error {
-	am, err := l.GetAccountMeta(address, txns...)
-	if err != nil {
-		return err
-	}
-	tokens := am.Tokens
-	for index, token := range am.Tokens {
-		if token.Type == meta.Type {
-			am.Tokens = append(tokens[:index], tokens[index+1:]...)
-			am.Tokens = append(am.Tokens, meta)
-			return l.UpdateAccountMeta(am, txns...)
-		}
-	}
-
-	am.Tokens = append(am.Tokens, meta)
-	return l.UpdateAccountMeta(am, txns...)
-}
-
-func (l *Ledger) DeleteTokenMeta(address types.Address, tokenType types.Hash, txns ...db.StoreTxn) error {
-	am, err := l.GetAccountMetaConfirmed(address, txns...)
-	if err != nil {
-		return err
-	}
-	tokens := am.Tokens
-	for index, token := range tokens {
-		if token.Type == tokenType {
-			am.Tokens = append(tokens[:index], tokens[index+1:]...)
-		}
-	}
-	return l.UpdateAccountMeta(am, txns...)
-}
-
-func (l *Ledger) DeleteTokenMetaCache(address types.Address, tokenType types.Hash, txns ...db.StoreTxn) error {
-	am, err := l.GetAccountMetaCache(address, txns...)
-	if err != nil {
-		return err
-	}
-	tokens := am.Tokens
-	for index, token := range tokens {
-		if token.Type == tokenType {
-			am.Tokens = append(tokens[:index], tokens[index+1:]...)
-		}
-	}
-	return l.UpdateAccountMetaCache(am, txns...)
-}
-
-func (l *Ledger) HasTokenMeta(address types.Address, tokenType types.Hash, txns ...db.StoreTxn) (bool, error) {
-	am, err := l.GetAccountMeta(address, txns...)
-	if err != nil {
-		if err == ErrAccountNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	for _, t := range am.Tokens {
-		if t.Type == tokenType {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func (l *Ledger) AddRepresentation(key types.Address, diff *types.Benefit, txns ...db.StoreTxn) error {
-	spin := l.representCache.getLock(key.String())
-	spin.Lock()
-	defer spin.Unlock()
-
-	value, err := l.GetRepresentation(key, txns...)
-	if err != nil && err != ErrRepresentationNotFound {
-		l.logger.Errorf("getRepresentation error: %s ,address: %s", err, key)
-		return err
-	}
-
-	value.Balance = value.Balance.Add(diff.Balance)
-	value.Vote = value.Vote.Add(diff.Vote)
-	value.Network = value.Network.Add(diff.Network)
-	value.Oracle = value.Oracle.Add(diff.Oracle)
-	value.Storage = value.Storage.Add(diff.Storage)
-	value.Total = value.Total.Add(diff.Total)
-
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-	return l.representCache.updateMemory(key, value, txn)
-}
-
-func (l *Ledger) SubRepresentation(key types.Address, diff *types.Benefit, txns ...db.StoreTxn) error {
-	spin := l.representCache.getLock(key.String())
-	spin.Lock()
-	defer spin.Unlock()
-
-	value, err := l.GetRepresentation(key, txns...)
-	if err != nil {
-		l.logger.Errorf("GetRepresentation error: %s ,address: %s", err, key)
-		return err
-	}
-	value.Balance = value.Balance.Sub(diff.Balance)
-	value.Vote = value.Vote.Sub(diff.Vote)
-	value.Network = value.Network.Sub(diff.Network)
-	value.Oracle = value.Oracle.Sub(diff.Oracle)
-	value.Storage = value.Storage.Sub(diff.Storage)
-	value.Total = value.Total.Sub(diff.Total)
-
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-	return l.representCache.updateMemory(key, value, txn)
-}
-
-func (l *Ledger) getRepresentation(key types.Address, txns ...db.StoreTxn) (*types.Benefit, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixRepresentation, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.Benefit)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			l.logger.Errorf("Unmarshal benefit error: %s ,address: %s, val: %s", err, key, string(v))
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrRepresentationNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetRepresentation(key types.Address, txns ...db.StoreTxn) (*types.Benefit, error) {
-	if lr, ok := l.representCache.getFromMemory(key.String()); ok {
-		return lr.(*types.Benefit), nil
-	}
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixRepresentation, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.Benefit)
-	err = txn.Get(k, func(v []byte, b byte) (err error) {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return &types.Benefit{
-				Vote:    types.ZeroBalance,
-				Network: types.ZeroBalance,
-				Storage: types.ZeroBalance,
-				Oracle:  types.ZeroBalance,
-				Balance: types.ZeroBalance,
-				Total:   types.ZeroBalance,
-			}, ErrRepresentationNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetRepresentations(fn func(types.Address, *types.Benefit) error, txns ...db.StoreTxn) error {
-	err := l.representCache.iterMemory(func(s string, i interface{}) error {
-		address, err := types.HexToAddress(s)
-		if err != nil {
-			return err
-		}
-		benefit := i.(*types.Benefit)
-		if err := fn(address, benefit); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-	err = txn.Iterator(idPrefixRepresentation, func(key []byte, val []byte, b byte) error {
-		address, err := types.BytesToAddress(key[1:])
-		if err != nil {
-			return err
-		}
-		if _, ok := l.representCache.getFromMemory(address.String()); !ok {
-			benefit := new(types.Benefit)
-			if err = benefit.Deserialize(val); err != nil {
-				l.logger.Errorf("Unmarshal benefit error: %s ,val: %s", err, string(val))
-				return err
-			}
-			if err := fn(address, benefit); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) GetRepresentationsCache(address types.Address, fn func(address types.Address, am *types.Benefit, amCache *types.Benefit) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	if !address.IsZero() {
-		be, err := l.getRepresentation(address)
-		if err != nil && err != ErrRepresentationNotFound {
-			return err
-		}
-		if v, ok := l.representCache.getFromMemory(address.String()); ok {
-			beMemory := v.(*types.Benefit)
-			if err := fn(address, be, beMemory); err != nil {
-				return err
-			}
-		} else {
-			if err := fn(address, be, nil); err != nil {
-				return err
-			}
-		}
-		return nil
-	} else {
-		err := l.representCache.iterMemory(func(s string, i interface{}) error {
-			addr, err := types.HexToAddress(s)
-			if err != nil {
-				return err
-			}
-			beMemory := i.(*types.Benefit)
-			be, err := l.getRepresentation(addr)
-			if err != nil && err != ErrRepresentationNotFound {
-				return err
-			}
-			if err := fn(addr, be, beMemory); err != nil {
-				return err
-			}
-			return nil
-		})
-		if err != nil {
-			return err
-		}
-
-		err = txn.Iterator(idPrefixRepresentation, func(key []byte, val []byte, b byte) error {
-			addr, err := types.BytesToAddress(key[1:])
-			if err != nil {
-				l.logger.Error(err)
-				return err
-			}
-			be := new(types.Benefit)
-			_, err = be.UnmarshalMsg(val)
-			if err != nil {
-				return err
-			}
-			if _, ok := l.representCache.getFromMemory(addr.String()); !ok {
-				if err := fn(addr, be, nil); err != nil {
-					return err
-				}
-			}
-			return nil
-		})
-
-		if err != nil {
-			return err
-		}
-		return nil
-	}
-}
-
-func (l *Ledger) AddPending(key *types.PendingKey, value *types.PendingInfo, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPending, key)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrPendingExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-
-	return txn.SetWithMeta(k, v, byte(types.PendingNotUsed))
-	//return txn.Set(key, pendingBytes)
-}
-
-func (l *Ledger) UpdatePending(key *types.PendingKey, kind types.PendingKind, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPending, key)
-	if err != nil {
-		return err
-	}
-	var pending types.PendingInfo
-	err = txn.Get(k[:], func(v []byte, b byte) (err error) {
-		if err := pending.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil
-		}
-		return err
-	}
-
-	v, err := pending.MarshalMsg(nil)
-	if err != nil {
-		return err
-	}
-
-	return txn.SetWithMeta(k, v, byte(kind))
-}
-
-func (l *Ledger) GetPending(key *types.PendingKey, txns ...db.StoreTxn) (*types.PendingInfo, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPending, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.PendingInfo)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrPendingNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) GetPendings(fn func(pendingKey *types.PendingKey, pendingInfo *types.PendingInfo) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	errStr := make([]string, 0)
-	err := txn.Iterator(idPrefixPending, func(key []byte, val []byte, b byte) error {
-		pendingKey := new(types.PendingKey)
-		if err := pendingKey.Deserialize(key[1:]); err != nil {
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		pendingInfo := new(types.PendingInfo)
-		if err := pendingInfo.Deserialize(val); err != nil {
-			errStr = append(errStr, err.Error())
-			return nil
-		}
-		used := types.PendingKind(b)
-		if used == types.PendingNotUsed {
-			if err := fn(pendingKey, pendingInfo); err != nil {
-				l.logger.Error("process pending error %s", err)
-				errStr = append(errStr, err.Error())
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	if len(errStr) != 0 {
-		return errors.New(strings.Join(errStr, ", "))
-	}
-	return nil
-}
-
-func (l *Ledger) SearchPending(address types.Address, fn func(key *types.PendingKey, value *types.PendingInfo) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Stream([]byte{idPrefixPending}, func(item *badger.Item) bool {
-		key := &types.PendingKey{}
-		if err := key.Deserialize(item.Key()[1:]); err == nil {
-			return key.Address == address
-		} else {
-			l.logger.Error(util.ToString(key), err)
-		}
-
-		return false
-	}, func(list *pb.KVList) error {
-		for _, v := range list.Kv {
-			pk := &types.PendingKey{}
-			pi := &types.PendingInfo{}
-			used := types.PendingKind(v.UserMeta[0])
-
-			if used == types.PendingNotUsed {
-				if err := pk.Deserialize(v.Key[1:]); err != nil {
-					continue
-				}
-				if err := pi.Deserialize(v.Value); err != nil {
-					continue
-				}
-				err := fn(pk, pi)
-				if err != nil {
-					l.logger.Error(err)
-				}
-			}
-		}
-		return nil
-	})
-}
-
-func (l *Ledger) DeletePending(key *types.PendingKey, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPending, key)
-	if err != nil {
-		return err
-	}
-	return txn.Delete(k)
-}
-
-func (l *Ledger) AddFrontier(frontier *types.Frontier, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixFrontier, frontier.HeaderBlock)
-	if err != nil {
-		return err
-	}
-	v := frontier.OpenBlock[:]
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrFrontierExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) GetFrontier(key types.Hash, txns ...db.StoreTxn) (*types.Frontier, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixFrontier, key)
-	if err != nil {
-		return nil, err
-	}
-
-	frontier := types.Frontier{HeaderBlock: key}
-	err = txn.Get(k, func(v []byte, b byte) (err error) {
-		copy(frontier.OpenBlock[:], v)
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrFrontierNotFound
-		}
-		return nil, err
-	}
-	return &frontier, nil
-}
-
-func (l *Ledger) GetFrontiers(txns ...db.StoreTxn) ([]*types.Frontier, error) {
-	var frontiers []*types.Frontier
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixFrontier, func(key []byte, val []byte, b byte) error {
-		var frontier types.Frontier
-		copy(frontier.HeaderBlock[:], key[1:])
-		copy(frontier.OpenBlock[:], val)
-		frontiers = append(frontiers, &frontier)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	sort.Sort(types.Frontiers(frontiers))
-	return frontiers, nil
-}
-
-func (l *Ledger) DeleteFrontier(key types.Hash, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixFrontier, key)
-	if err != nil {
-		return err
-	}
-	return txn.Delete(k)
-}
-
-func (l *Ledger) CountFrontiers(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Count([]byte{idPrefixFrontier})
-}
-
-func (l *Ledger) GetChild(key types.Hash, txns ...db.StoreTxn) (types.Hash, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixChild, key)
-	if err != nil {
-		return types.ZeroHash, err
-	}
-
-	value := new(types.Hash)
-	err = txn.Get(k, func(val []byte, b byte) error {
-		if _, err := value.UnmarshalMsg(val); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return types.ZeroHash, err
-	}
-	return *value, nil
-}
-
-func (l *Ledger) GetLinkBlock(key types.Hash, txns ...db.StoreTxn) (types.Hash, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixLink, key)
-	if err != nil {
-		return types.ZeroHash, err
-	}
-	value := new(types.Hash)
-	err = txn.Get(k, func(val []byte, b byte) (err error) {
-		if _, err := value.UnmarshalMsg(val); err != nil {
-			return errors.New("unmarshal hash error")
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return types.ZeroHash, ErrLinkNotFound
-		}
-		return types.ZeroHash, fmt.Errorf("get link error: %s", err)
-	}
-	return *value, nil
-}
-
 func getVersionKey() []byte {
 	return []byte{idPrefixVersion}
 }
@@ -1756,377 +361,6 @@ func getVersion(txn db.StoreTxn) (int64, error) {
 		return i, err
 	}
 	return i, nil
-}
-
-//getTxn get txn by `update` mode
-func (l *Ledger) getTxn(update bool, txns ...db.StoreTxn) (db.StoreTxn, bool) {
-	if len(txns) > 0 {
-		//logger.Debugf("getTxn %p", txns[0])
-		return txns[0], false
-	} else {
-		txn := l.Store.NewTransaction(update)
-		//logger.Debugf("getTxn new %p", txn)
-		return txn, true
-	}
-}
-
-// releaseTxn commit change and close txn
-func (l *Ledger) releaseTxn(txn db.StoreTxn, flag bool) {
-	if flag {
-		err := txn.Commit(nil)
-		if err != nil {
-			l.logger.Error(err)
-		}
-		txn.Discard()
-	}
-}
-
-// BatchUpdate MUST pass the same txn
-func (l *Ledger) BatchUpdate(fn func(txn db.StoreTxn) error) error {
-	txn := l.Store.NewTransaction(true)
-	//logger.Debugf("BatchUpdate NewTransaction %p", txn)
-	defer func() {
-		//logger.Debugf("BatchUpdate Discard %p", txn)
-		txn.Discard()
-	}()
-
-	if err := fn(txn); err != nil {
-		return err
-	}
-	return txn.Commit(nil)
-}
-
-// BatchView MUST pass the same txn
-func (l *Ledger) BatchView(fn func(txn db.StoreTxn) error) error {
-	txn := l.Store.NewTransaction(false)
-	//logger.Debugf("BatchView NewTransaction %p", txn)
-	defer func() {
-		//logger.Debugf("BatchView Discard %p", txn)
-		txn.Discard()
-	}()
-
-	if err := fn(txn); err != nil {
-		return err
-	}
-	return txn.Commit(nil)
-}
-
-func (l *Ledger) Latest(account types.Address, token types.Hash, txns ...db.StoreTxn) types.Hash {
-	zero := types.Hash{}
-	am, err := l.GetAccountMeta(account, txns...)
-	if err != nil {
-		return zero
-	}
-	tm := am.Token(token)
-	if tm != nil {
-		return tm.Header
-	}
-	return zero
-}
-
-func (l *Ledger) Account(hash types.Hash, txns ...db.StoreTxn) (*types.AccountMeta, error) {
-	block, err := l.GetStateBlock(hash, txns...)
-	if err != nil {
-		return nil, err
-	}
-	addr := block.GetAddress()
-	am, err := l.GetAccountMeta(addr, txns...)
-	if err != nil {
-		return nil, err
-	}
-
-	return am, nil
-}
-
-func (l *Ledger) Token(hash types.Hash, txns ...db.StoreTxn) (*types.TokenMeta, error) {
-	block, err := l.GetStateBlockConfirmed(hash, txns...)
-	if err != nil {
-		return nil, err
-	}
-	token := block.GetToken()
-	addr := block.GetAddress()
-	am, err := l.GetAccountMetaConfirmed(addr, txns...)
-	if err != nil {
-		return nil, err
-	}
-
-	tm := am.Token(token)
-	if tm != nil {
-		return tm, nil
-	}
-
-	//TODO: hash to token name
-	return nil, fmt.Errorf("can not find token %s", token)
-}
-
-func (l *Ledger) Pending(account types.Address, txns ...db.StoreTxn) ([]*types.PendingKey, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	var cache []*types.PendingKey
-	err := txn.Iterator(idPrefixPending, func(key []byte, val []byte, b byte) error {
-		pendingKey := types.PendingKey{}
-		_, err := pendingKey.UnmarshalMsg(key[1:])
-		if err != nil {
-			return err
-		}
-		if pendingKey.Address == account {
-			cache = append(cache, &pendingKey)
-		}
-		return nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return cache, nil
-}
-
-func (l *Ledger) Balance(account types.Address, txns ...db.StoreTxn) (map[types.Hash]types.Balance, error) {
-	cache := make(map[types.Hash]types.Balance)
-	am, err := l.GetAccountMeta(account, txns...)
-	if err != nil {
-		return cache, err
-	}
-	for _, tm := range am.Tokens {
-		cache[tm.Type] = tm.Balance
-	}
-
-	if len(cache) == 0 {
-		return nil, fmt.Errorf("can not find any token balance ")
-	}
-
-	return cache, nil
-}
-
-func (l *Ledger) TokenPending(account types.Address, token types.Hash, txns ...db.StoreTxn) ([]*types.PendingKey, error) {
-	var cache []*types.PendingKey
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixPending, func(key []byte, val []byte, b byte) error {
-		pendingKey := types.PendingKey{}
-		_, err := pendingKey.UnmarshalMsg(key[1:])
-		if err != nil {
-			return nil
-		}
-		if pendingKey.Address == account {
-			var pending types.PendingInfo
-			_, err := pending.UnmarshalMsg(val)
-			if err != nil {
-				return err
-			}
-			if pending.Type == token {
-				cache = append(cache, &pendingKey)
-			}
-		}
-		return nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return cache, nil
-}
-
-func (l *Ledger) TokenPendingInfo(account types.Address, token types.Hash, txns ...db.StoreTxn) ([]*types.PendingInfo, error) {
-	var cache []*types.PendingInfo
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixPending, func(key []byte, val []byte, b byte) error {
-		pendingKey := types.PendingKey{}
-		_, err := pendingKey.UnmarshalMsg(key[1:])
-		if err != nil {
-			return nil
-		}
-		if pendingKey.Address == account {
-			var pendingInfo types.PendingInfo
-			_, err := pendingInfo.UnmarshalMsg(val)
-			if err != nil {
-				return err
-			}
-			if pendingInfo.Type == token {
-				cache = append(cache, &pendingInfo)
-			}
-		}
-		return nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return cache, nil
-}
-
-func (l *Ledger) TokenBalance(account types.Address, token types.Hash, txns ...db.StoreTxn) (types.Balance, error) {
-	am, err := l.GetAccountMeta(account, txns...)
-	if err != nil {
-		return types.ZeroBalance, err
-	}
-	tm := am.Token(token)
-	if tm != nil {
-		return tm.Balance, nil
-	}
-
-	return types.ZeroBalance, fmt.Errorf("can not find %s balance", token)
-}
-
-func (l *Ledger) Weight(account types.Address, txns ...db.StoreTxn) types.Balance {
-	benefit, err := l.GetRepresentation(account, txns...)
-	if err != nil {
-		return types.ZeroBalance
-	}
-	return benefit.Total
-}
-
-func (l *Ledger) AddOrUpdatePerformance(value *types.PerformanceTime, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPerformance, value.Hash)
-	if err != nil {
-		return err
-	}
-	v, err := json.Marshal(value)
-	if err != nil {
-		return err
-	}
-
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) PerformanceTimes(fn func(*types.PerformanceTime), txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixPerformance, func(key []byte, val []byte, b byte) error {
-		pt := new(types.PerformanceTime)
-		err := json.Unmarshal(val, pt)
-		if err != nil {
-			return err
-		}
-		fn(pt)
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) GetPerformanceTime(key types.Hash, txns ...db.StoreTxn) (*types.PerformanceTime, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixPerformance, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := types.NewPerformanceTime()
-	err = txn.Get(k, func(val []byte, b byte) (err error) {
-		return json.Unmarshal(val, &value)
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrPerformanceNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) IsPerformanceTimeExist(key types.Hash, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	if _, err := l.GetPerformanceTime(key); err == nil {
-		return true, nil
-	} else {
-		if err == ErrPerformanceNotFound {
-			return false, nil
-		} else {
-			return false, err
-		}
-	}
-}
-
-func (l *Ledger) CalculateAmount(block *types.StateBlock, txns ...db.StoreTxn) (types.Balance, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	var prev *types.StateBlock
-	var err error
-	switch block.GetType() {
-	case types.Open:
-		return block.TotalBalance(), err
-	case types.Send:
-		if prev, err = l.GetStateBlock(block.GetPrevious(), txn); err != nil {
-			return types.ZeroBalance, err
-		}
-		return prev.TotalBalance().Sub(block.TotalBalance()), nil
-	case types.Receive:
-		if prev, err = l.GetStateBlock(block.GetPrevious(), txn); err != nil {
-			return types.ZeroBalance, err
-		}
-		return block.TotalBalance().Sub(prev.TotalBalance()), nil
-	case types.Change:
-		return types.ZeroBalance, nil
-	case types.ContractReward:
-		prevHash := block.GetPrevious()
-		if prevHash.IsZero() {
-			return block.TotalBalance(), nil
-		} else {
-			if prev, err = l.GetStateBlock(prevHash, txn); err != nil {
-				return types.ZeroBalance, err
-			}
-			return block.TotalBalance().Sub(prev.TotalBalance()), nil
-		}
-	case types.ContractSend:
-		if common.IsGenesisBlock(block) {
-			return block.GetBalance(), nil
-		} else {
-			if prev, err = l.GetStateBlock(block.GetPrevious(), txn); err != nil {
-				return types.ZeroBalance, err
-			}
-			return prev.TotalBalance().Sub(block.TotalBalance()), nil
-		}
-	default:
-		return types.ZeroBalance, errors.New("invalid block type")
-	}
-}
-
-func (l *Ledger) GetGenesis(txns ...db.StoreTxn) ([]*types.StateBlock, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	var blocks []*types.StateBlock
-	err := txn.Iterator(idPrefixToken, func(key []byte, val []byte, b byte) error {
-		t := new(types.TokenInfo)
-		if err := json.Unmarshal(val, t); err != nil {
-			return err
-		}
-		tm, err := l.GetTokenMeta(t.Owner, t.TokenId, txn)
-		if err != nil {
-			return err
-		}
-		block, err := l.GetStateBlock(tm.OpenBlock, txn)
-		if err != nil {
-			return err
-		}
-		blocks = append(blocks, block)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return blocks, nil
 }
 
 func (l *Ledger) generateWork(hash types.Hash) types.Work {
@@ -2293,311 +527,6 @@ func (l *Ledger) GenerateChangeBlock(account types.Address, representative types
 		sb.Work = l.generateWork(sb.Root())
 	}
 	return &sb, nil
-}
-
-func (l *Ledger) GetOnlineRepresentations(txns ...db.StoreTxn) ([]types.Address, error) {
-	key := []byte{idPrefixOnlineReps}
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-	var result []types.Address
-	err := txn.Get(key, func(bytes []byte, b byte) error {
-		if len(bytes) > 0 {
-			return json.Unmarshal(bytes, &result)
-		}
-		return nil
-	})
-
-	return result, err
-}
-
-func (l *Ledger) SetOnlineRepresentations(addresses []*types.Address, txns ...db.StoreTxn) error {
-	bytes, err := json.Marshal(addresses)
-	if err != nil {
-		return err
-	}
-
-	key := []byte{idPrefixOnlineReps}
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Set(key, bytes)
-}
-
-func (l *Ledger) AddMessageInfo(key types.Hash, value []byte, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixMessageInfo, key)
-	if err != nil {
-		return err
-	}
-	if err := txn.Set(k, value); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) GetMessageInfo(key types.Hash, txns ...db.StoreTxn) ([]byte, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixMessageInfo, key)
-	if err != nil {
-		return nil, err
-	}
-	var value []byte
-	err = txn.Get(k, func(v []byte, b byte) error {
-		value = v
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) AddBlockCache(value *types.StateBlock, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCache, value.GetHash())
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrBlockExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) GetBlockCache(key types.Hash, txns ...db.StoreTxn) (*types.StateBlock, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCache, key)
-	if err != nil {
-		return nil, err
-	}
-
-	value := new(types.StateBlock)
-	err = txn.Get(k, func(v []byte, b byte) error {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrBlockNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) HasBlockCache(key types.Hash, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCache, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
-}
-
-func (l *Ledger) DeleteBlockCache(key types.Hash, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCache, key)
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	if err := txn.Delete(k); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (l *Ledger) CountBlockCache(txns ...db.StoreTxn) (uint64, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	return txn.Count([]byte{idPrefixBlockCache})
-}
-
-func (l *Ledger) GetBlockCaches(fn func(*types.StateBlock) error, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	err := txn.Iterator(idPrefixBlockCache, func(key []byte, val []byte, b byte) error {
-		blk := new(types.StateBlock)
-		if err := blk.Deserialize(val); err != nil {
-			l.logger.Errorf("deserialize block error: %s", err)
-			return nil
-		}
-		if err := fn(blk); err != nil {
-			l.logger.Errorf("process block error: %s", err)
-			return nil
-		}
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (l *Ledger) AddAccountMetaCache(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(v []byte, b byte) error {
-		return nil
-	})
-	if err == nil {
-		return ErrAccountExists
-	} else if err != badger.ErrKeyNotFound {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) GetAccountMetaCache(key types.Address, txns ...db.StoreTxn) (*types.AccountMeta, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, key)
-	if err != nil {
-		return nil, err
-	}
-	value := new(types.AccountMeta)
-	err = txn.Get(k, func(v []byte, b byte) (err error) {
-		if err := value.Deserialize(v); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return nil, ErrAccountNotFound
-		}
-		return nil, err
-	}
-	return value, nil
-}
-
-func (l *Ledger) AddOrUpdateAccountMetaCache(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) UpdateAccountMetaCache(value *types.AccountMeta, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, value.Address)
-	if err != nil {
-		return err
-	}
-	v, err := value.Serialize()
-	if err != nil {
-		return err
-	}
-
-	err = txn.Get(k, func(vals []byte, b byte) error {
-		return nil
-	})
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return ErrAccountNotFound
-		}
-		return err
-	}
-	return txn.Set(k, v)
-}
-
-func (l *Ledger) DeleteAccountMetaCache(key types.Address, txns ...db.StoreTxn) error {
-	txn, flag := l.getTxn(true, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, key)
-	if err != nil {
-		return err
-	}
-
-	return txn.Delete(k)
-}
-
-func (l *Ledger) HasAccountMetaCache(key types.Address, txns ...db.StoreTxn) (bool, error) {
-	txn, flag := l.getTxn(false, txns...)
-	defer l.releaseTxn(txn, flag)
-
-	k, err := getKeyOfParts(idPrefixBlockCacheAccount, key)
-	if err != nil {
-		return false, err
-	}
-	err = txn.Get(k, func(val []byte, b byte) error {
-		return nil
-	})
-
-	if err != nil {
-		if err == badger.ErrKeyNotFound {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
 }
 
 func (l *Ledger) BlockCacheProcess(block types.Block) error {
