+++ ledger/ledger_frontier.go
@@ -1,0 +1,90 @@
+package ledger
+
+import (
+	"sort"
+
+	"github.com/dgraph-io/badger"
+	"github.com/qlcchain/go-qlc/common/types"
+	"github.com/qlcchain/go-qlc/ledger/db"
+)
+
+func (l *Ledger) AddFrontier(frontier *types.Frontier, txns ...db.StoreTxn) error {
+	txn, flag := l.getTxn(true, txns...)
+	defer l.releaseTxn(txn, flag)
+
+	k, err := getKeyOfParts(idPrefixFrontier, frontier.HeaderBlock)
+	if err != nil {
+		return err
+	}
+	v := frontier.OpenBlock[:]
+
+	err = txn.Get(k, func(v []byte, b byte) error {
+		return nil
+	})
+	if err == nil {
+		return ErrFrontierExists
+	} else if err != badger.ErrKeyNotFound {
+		return err
+	}
+	return txn.Set(k, v)
+}
+
+func (l *Ledger) GetFrontier(key types.Hash, txns ...db.StoreTxn) (*types.Frontier, error) {
+	txn, flag := l.getTxn(false, txns...)
+	defer l.releaseTxn(txn, flag)
+
+	k, err := getKeyOfParts(idPrefixFrontier, key)
+	if err != nil {
+		return nil, err
+	}
+
+	frontier := types.Frontier{HeaderBlock: key}
+	err = txn.Get(k, func(v []byte, b byte) (err error) {
+		copy(frontier.OpenBlock[:], v)
+		return nil
+	})
+	if err != nil {
+		if err == badger.ErrKeyNotFound {
+			return nil, ErrFrontierNotFound
+		}
+		return nil, err
+	}
+	return &frontier, nil
+}
+
+func (l *Ledger) GetFrontiers(txns ...db.StoreTxn) ([]*types.Frontier, error) {
+	var frontiers []*types.Frontier
+	txn, flag := l.getTxn(false, txns...)
+	defer l.releaseTxn(txn, flag)
+
+	err := txn.Iterator(idPrefixFrontier, func(key []byte, val []byte, b byte) error {
+		var frontier types.Frontier
+		copy(frontier.HeaderBlock[:], key[1:])
+		copy(frontier.OpenBlock[:], val)
+		frontiers = append(frontiers, &frontier)
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+	sort.Sort(types.Frontiers(frontiers))
+	return frontiers, nil
+}
+
+func (l *Ledger) DeleteFrontier(key types.Hash, txns ...db.StoreTxn) error {
+	txn, flag := l.getTxn(true, txns...)
+	defer l.releaseTxn(txn, flag)
+
+	k, err := getKeyOfParts(idPrefixFrontier, key)
+	if err != nil {
+		return err
+	}
+	return txn.Delete(k)
+}
+
+func (l *Ledger) CountFrontiers(txns ...db.StoreTxn) (uint64, error) {
+	txn, flag := l.getTxn(false, txns...)
+	defer l.releaseTxn(txn, flag)
+
+	return txn.Count([]byte{idPrefixFrontier})
+}
