+++ ledger/ledger_represention_test.go
@@ -1,0 +1,165 @@
+package ledger
+
+import (
+	"math/big"
+	"testing"
+
+	"github.com/qlcchain/go-qlc/common/types"
+	"github.com/qlcchain/go-qlc/mock"
+)
+
+func addRepresentationWeight(t *testing.T, l *Ledger) *types.AccountMeta {
+	address := mock.Address()
+	ac := mock.AccountMeta(address)
+	ac.CoinBalance = types.Balance{Int: big.NewInt(int64(1000))}
+	ac.CoinVote = types.Balance{Int: big.NewInt(int64(1000))}
+	benefit := &types.Benefit{
+		Vote:    ac.CoinVote,
+		Storage: ac.CoinStorage,
+		Network: ac.CoinNetwork,
+		Oracle:  ac.CoinOracle,
+		Balance: ac.CoinBalance,
+		Total:   ac.TotalBalance(),
+	}
+
+	err := l.AddRepresentation(address, benefit)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return ac
+}
+
+func TestLedger_AddRepresentationWeight(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+	ac := addRepresentationWeight(t, l)
+	diff := &types.Benefit{
+		Vote:    types.Balance{Int: big.NewInt(int64(10))},
+		Storage: types.ZeroBalance,
+		Network: types.ZeroBalance,
+		Oracle:  types.ZeroBalance,
+		Balance: types.Balance{Int: big.NewInt(int64(10))},
+		Total:   types.Balance{Int: big.NewInt(int64(20))},
+	}
+
+	err := l.AddRepresentation(ac.Address, diff)
+	if err != nil {
+		t.Fatal(err)
+	}
+	a, err := l.GetRepresentation(ac.Address)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Log(a)
+	if !a.Total.Equal(ac.TotalBalance().Add(diff.Total)) {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_SubRepresentationWeight(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	ac := addRepresentationWeight(t, l)
+	diff := &types.Benefit{
+		Vote:    types.Balance{Int: big.NewInt(int64(10))},
+		Storage: types.ZeroBalance,
+		Network: types.ZeroBalance,
+		Oracle:  types.ZeroBalance,
+		Balance: types.Balance{Int: big.NewInt(int64(10))},
+		Total:   types.Balance{Int: big.NewInt(int64(20))},
+	}
+
+	err := l.SubRepresentation(ac.Address, diff)
+	if err != nil {
+		t.Fatal(err)
+	}
+	a, err := l.GetRepresentation(ac.Address)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Log(a)
+	if !a.Total.Equal(ac.TotalBalance().Sub(diff.Total)) {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_GetRepresentations(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	addRepresentationWeight(t, l)
+	addRepresentationWeight(t, l)
+
+	err := l.GetRepresentations(func(address types.Address, benefit *types.Benefit) error {
+		t.Log(address, benefit)
+		return nil
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_GetRepresentationsCache(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	addRepresentationWeight(t, l)
+	addRepresentationWeight(t, l)
+
+	err := l.GetRepresentationsCache(types.ZeroAddress, func(address types.Address, am *types.Benefit, amCache *types.Benefit) error {
+		t.Log(address, am, amCache)
+		return nil
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_OnlineRepresentations(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+	var addr []*types.Address
+	for i := 0; i < 10; i++ {
+		a1 := mock.Address()
+		addr = append(addr, &a1)
+	}
+
+	err := l.SetOnlineRepresentations(addr)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if addr2, err := l.GetOnlineRepresentations(); err == nil {
+		if len(addr) != len(addr2) || len(addr2) != 10 {
+			t.Fatal("invalid online rep size")
+		}
+		for i, v := range addr {
+			if v.String() != addr2[i].String() {
+				t.Fatal("invalid ")
+			} else {
+				t.Log(v.String())
+			}
+		}
+	} else {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_SetOnlineRepresentations(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	var addr []*types.Address
+	err := l.SetOnlineRepresentations(addr)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if addr2, err := l.GetOnlineRepresentations(); err == nil {
+		if len(addr2) != 0 {
+			t.Fatal("invalid online rep")
+		}
+	} else {
+		t.Fatal(err)
+	}
+}
