+++ ledger/ledger_uncheck_test.go
@@ -1,0 +1,94 @@
+package ledger
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/qlcchain/go-qlc/common/types"
+	"github.com/qlcchain/go-qlc/mock"
+)
+
+func addUncheckedBlock(t *testing.T, l *Ledger) (hash types.Hash, block *types.StateBlock, kind types.UncheckedKind) {
+	block = mock.StateBlockWithoutWork()
+	hash = block.GetLink()
+	kind = types.UncheckedKindLink
+	fmt.Println(hash)
+	if err := l.AddUncheckedBlock(hash, block, kind, types.UnSynchronized); err != nil {
+		t.Fatal(err)
+	}
+	return
+}
+
+func TestLedger_AddUncheckedBlock(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+	addUncheckedBlock(t, l)
+}
+
+func TestLedger_GetUncheckedBlock(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	parentHash, _, kind := addUncheckedBlock(t, l)
+
+	if b, s, err := l.GetUncheckedBlock(parentHash, kind); err != nil {
+		t.Fatal(err)
+	} else {
+		t.Logf("unchecked,%s", b)
+		t.Log(s)
+	}
+
+}
+
+func TestLedger_CountUncheckedBlocks(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+	addUncheckedBlock(t, l)
+	addUncheckedBlock(t, l)
+
+	c, err := l.CountUncheckedBlocks()
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Log("unchecked count,", c)
+}
+
+func TestLedger_HasUncheckedBlock(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	parentHash, _, kind := addUncheckedBlock(t, l)
+	r, err := l.HasUncheckedBlock(parentHash, kind)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Log("has unchecked,", r)
+}
+
+func TestLedger_GetUncheckedBlocks(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+	addUncheckedBlock(t, l)
+	addUncheckedBlock(t, l)
+
+	err := l.WalkUncheckedBlocks(func(block *types.StateBlock, link types.Hash, unCheckType types.UncheckedKind, sync types.SynchronizedKind) error {
+		t.Log(block)
+		t.Log(link, unCheckType, sync)
+
+		return nil
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestLedger_DeleteUncheckedBlock(t *testing.T) {
+	teardownTestCase, l := setupTestCase(t)
+	defer teardownTestCase(t)
+
+	parentHash, _, kind := addUncheckedBlock(t, l)
+	err := l.DeleteUncheckedBlock(parentHash, kind)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
